removing old files and renaming files
