--- src/utils.ts
@@ -1,3 +1,0 @@
-export function isString(obj: unknown): obj is string {
-  return Object.prototype.toString.call(obj) === '[object String]'
-}
